Remove unnecessary casts

+++ common/src/commonMain/kotlin/cz/frantisekmasa/wfrp_master/common/core/domain/character/Character.kt
@@ -265,10 +265,10 @@ data class Character(
             )
             return (
                 baseWounds +
-                    modifiers.extraToughnessBonusMultiplier.toInt() * toughnessBonus +
+                    modifiers.extraToughnessBonusMultiplier * toughnessBonus +
                     (if (hasHardyTalent) toughnessBonus else 0)
                 ) *
-                modifiers.afterMultiplier.toInt()
+                modifiers.afterMultiplier
         }
     }
 }
